20250330添加数据查询 编译完成 未调试

+++ SoftWare/中显科技组态屏/STM32MCU/Dev/WeightMode_app/Inc/app_AT24C.h
@@ -6,7 +6,7 @@
 
 
 typedef uint8   ExtFlashDevAddType;
-typedef uint16  ExtFlashRegAddType;
+typedef uint32  ExtFlashRegAddType;
 typedef uint8   ExtFlashRegVluType;
 
 typedef enum
